requirements.txt
alembic==1.8.1
asgiref==3.5.0
astroid==2.11.2
Authlib==0.15.5
azure-common==1.1.28
azure-core==1.26.0
azure-identity==1.9.0
azure-mgmt-core==1.3.2
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==21.0.0
azure-mgmt-subscription==3.0.0
Babel==2.10.3
bcrypt==3.2.2
bidict==0.22.0
blinker==1.5
boto3==1.20.54
botocore==1.23.54
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
click==8.1.3
cryptography==3.4.8
dill==0.3.4
Django==4.0.3
django-cleanup==6.0.0
django-phonenumber-field==6.1.0
django-recaptcha==3.0.0
django-storages==1.13.1
dnspython==2.2.1
email-validator==1.3.0
eventlet==0.33.0
Flask==2.1.3
Flask-Babel==2.0.0
Flask-Compress==1.13
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.5
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
gunicorn==20.1.0
httpagentparser==1.9.3
idna==2.10
isodate==0.6.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
lazy-object-proxy==1.7.1
ldap3==2.9.1
Mako==1.2.3
MarkupSafe==2.1.1
mccabe==0.7.0
msal==1.19.0
msal-extensions==0.3.1
msrest==0.7.1
oauthlib==3.2.1
paramiko==2.11.0
passlib==1.7.4
pgadmin4==6.14
phonenumbers==8.12.45
Pillow==9.0.1
platformdirs==2.5.1
portalocker==2.5.1
psutil==5.9.2
psycopg2==2.9.4
psycopg2-binary==2.9.4
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.5.0
pylint==2.13.3
PyNaCl==1.5.0
pyotp==2.7.0
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2021.3
qrcode==7.3.1
requests==2.25.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
simplejson==3.17.6
six==1.16.0
speaklater3==1.4
SQLAlchemy==1.4.41
sqlparse==0.4.2
sshtunnel==0.4.0
tomli==2.0.1
types-cryptography==3.3.23
typing_extensions==4.3.0
ua-parser==0.16.1
urllib3==1.26.12
user-agents==2.2.0
Werkzeug==2.1.2
whitenoise==6.2.0
wrapt==1.14.0
WTForms==3.0.1
